Unique results
Add a checkbox to list unique results only.

Unique in this context means files with the same BHash, that is,
the same file content (files can have the same content but different
paths in the filesystem).

The default behavior is to list duplicated files by default, following
the principle of least surprise, to prevent hiding relevant info to the
user, as there could be interest in listing where the file is duplicated
in a backup.